pyproject.toml
[tool.poetry]
name = "task-management-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.8.0"
django = "3.2.0"
dependency-injector = "^4.41.0"
pydantic = "^2.7.4"
djangorestframework = "^3.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
